+++ src/Stratis.Features.FederatedPeg/Interfaces/IInputConsolidator.cs
@@ -4,6 +4,7 @@ using System.Text;
 using NBitcoin;
 using Stratis.Bitcoin.Primitives;
 using Stratis.Features.FederatedPeg.Events;
+using Stratis.Features.FederatedPeg.InputConsolidation;
 using Stratis.Features.FederatedPeg.TargetChain;
 
 namespace Stratis.Features.FederatedPeg.Interfaces
