+++ src/Stratis.Features.FederatedPeg/Payloads/RequestPartialTransactionPayload.cs
@@ -1,5 +1,6 @@
 ﻿using NBitcoin;
 using Stratis.Bitcoin.P2P.Protocol.Payloads;
+using Stratis.Features.FederatedPeg.InputConsolidation;
 using Stratis.Features.FederatedPeg.TargetChain;
 
 namespace Stratis.Features.FederatedPeg.Payloads
